+++ app/components/home/WhatToExpectDesktop.tsx
@@ -155,8 +155,8 @@ export default function WhatToExpectDesktop({
               <Image
                 src={
                   selectedFormula === "01"
-                    ? "/lifestyle/FlowBoxOpen.jpg"
-                    : "/lifestyle/ClearLaugh.jpg"
+                    ? "/lifestyle/flow/FlowBoxOpen.jpg"
+                    : "/lifestyle/clear/ClearLaugh.jpg"
                 }
                 alt={
                   selectedFormula === "01"